back to an old version
new version was having trouble compiling the .bib files (it was hanging when it got to them).

I don't know what the problem was, but I suspect it was introduced in a new version of pandoc-citeproc or something. Also it could have something to do with macOS's new weird security privileges, since the error didn't show up on the linux machine in github actions.